:sparkles: v4.0 :sparkles:
Complete redesign of the site.

Also tweaked the /frames folder around a bit, the framework relied on
bootstrap v3.x and the new design works with v4, for this reason I
added the bootstrap v3 to the assets folder of /frames.

This is a bit of a double dependency, but oh well. It works.