Fix weird white bar.
When playing the initial load in animation the area behind the header
background image was displaying white.

This was caused in one of the previous commits where we moved the
background-color from the body to the .content for some reason.

Turns out that this change was not needed really; the body could just
have a background-color… In this commit we add back the background
color to the body to make sure we don’t get any “ugly” (out of theme)
white bars while loading.