:sparkles: v5.0.1 :sparkles:
This update introduces some small design changes. Primarily the fact
that we took the paragraph styles from the ‘my story’ section and used
it to group the links and info in the ‘info’ section a little bit
better.

This also gives the site a more uniform feel and shows off our accent
colour (currently blue) a little bit better.