Introducing scroll lock
When the mobile navigation was open you could still scroll up and down
the page. This wasn’t the expected result.

I had to build some safe guards to avoid awkward situations where the
user toggles between the large and the small version of the navbar.
(The large version is always at the top of the document. The small one
scrolls with.)

The scrips look like a mess in my opinion… but they work!