Font weights and Header animation
Added a 3th font weight to make readability better. The paragraphs at
`font-weight: 100` we barely readable especially on mobile devices.
However I did want to keep some of the headlines in this thinner font
weight and some headlines in the bolder 500 font weight. So yeah. I
introduced the 300 version for the plain text.

Besides that I added a fancy background color breathing animation to
the background of the two words in the header that are emphasised.
Makes them stand out more.